+++ AmplifyPlugins/DataStore/Tests/DataStoreHostApp/AWSDataStorePluginLazyLoadTests/LL14/AWSDataStoreLazyLoadUserPostCommentSnapshotTests.swift
@@ -11,6 +11,3 @@ import XCTest
 
 @testable import Amplify
 
-extension AWSDataStoreLazyLoadUserPostCommentTests {
-
-}