+++ Package.swift
@@ -9,7 +9,7 @@ let platforms: [SupportedPlatform] = [
     .watchOS(.v9)
 ]
 let dependencies: [Package.Dependency] = [
-    .package(url: "https://github.com/awslabs/aws-sdk-swift", exact: "1.6.7"),
+    .package(url: "https://github.com/awslabs/aws-sdk-swift", exact: "1.6.71"),
     .package(url: "https://github.com/stephencelis/SQLite.swift.git", exact: "0.15.3"),
     .package(url: "https://github.com/mattgallagher/CwlPreconditionTesting.git", from: "2.1.0"),
     .package(url: "https://github.com/aws-amplify/amplify-swift-utils-notifications.git", from: "1.1.0")
@@ -367,6 +367,32 @@ let analyticsTargets: [Target] = [
             "AmplifyTestCommon"
         ],
         path: "AmplifyPlugins/Analytics/Tests/AWSPinpointAnalyticsPluginUnitTests"
+    ),
+]
+
+let kinesisTargets: [Target] = [
+    .target(
+        name: "AmplifyKinesisClient",
+        dependencies: [
+            .target(name: "AmplifyFoundation"),
+            .target(name: "AmplifyFoundationBridge"),
+            .product(name: "SQLite", package: "SQLite.swift"),
+            .product(name: "AWSKinesis", package: "aws-sdk-swift")
+        ],
+        path: "AmplifyClients/AmplifyKinesisClient/Sources",
+        resources: [
+            .copy("Resources/PrivacyInfo.xcprivacy")
+        ],
+        swiftSettings: [
+            .enableUpcomingFeature("StrictConcurrency")
+        ]
+    ),
+    .testTarget(
+        name: "AmplifyKinesisClientTests",
+        dependencies: [
+            "AmplifyKinesisClient"
+        ],
+        path: "AmplifyClients/AmplifyKinesisClient/Tests/UnitTests"
     )
 ]
 
@@ -470,17 +496,52 @@ let loggingTargets: [Target] = [
     )
 ]
 
-let targets: [Target] = amplifyTargets
-    + apiTargets
-    + authTargets
-    + dataStoreTargets
-    + storageTargets
-    + geoTargets
-    + analyticsTargets
-    + pushNotificationsTargets
-    + internalPinpointTargets
-    + predictionsTargets
-    + loggingTargets
+let foundationTargets: [Target] = [
+    .target(
+        name: "AmplifyFoundation",
+        dependencies: [],
+        path: "AmplifyFoundation/Sources"
+    ),
+    .testTarget(
+        name: "AmplifyFoundationTests",
+        dependencies: ["AmplifyFoundation"],
+        path: "AmplifyFoundation/Tests"
+    ),
+]
+
+let foundationBridgeTargets: [Target] = [
+    .target(
+        name: "AmplifyFoundationBridge",
+        dependencies: [
+            "AmplifyFoundation",
+            .product(name: "AWSClientRuntime", package: "aws-sdk-swift")
+        ],
+        path: "AmplifyFoundationBridge/Sources"
+    ),
+    .testTarget(
+        name: "AmplifyFoundationBridgeTests",
+        dependencies: [
+            "AmplifyFoundationBridge"
+        ],
+        path: "AmplifyFoundationBridge/Tests"
+    ),
+]
+
+var targets: [Target] = []
+targets.append(contentsOf: amplifyTargets)
+targets.append(contentsOf: apiTargets)
+targets.append(contentsOf: authTargets)
+targets.append(contentsOf: dataStoreTargets)
+targets.append(contentsOf: storageTargets)
+targets.append(contentsOf: geoTargets)
+targets.append(contentsOf: analyticsTargets)
+targets.append(contentsOf: kinesisTargets)
+targets.append(contentsOf: pushNotificationsTargets)
+targets.append(contentsOf: internalPinpointTargets)
+targets.append(contentsOf: predictionsTargets)
+targets.append(contentsOf: loggingTargets)
+targets.append(contentsOf: foundationTargets)
+targets.append(contentsOf: foundationBridgeTargets)
 
 let package = Package(
     name: "Amplify",
@@ -533,6 +594,18 @@ let package = Package(
         .library(
             name: "AWSCloudWatchLoggingPlugin",
             targets: ["AWSCloudWatchLoggingPlugin"]
+        ),
+        .library(
+            name: "AmplifyKinesisClient",
+            targets: ["AmplifyKinesisClient"]
+        ),
+        .library(
+            name: "AmplifyFoundation",
+            targets: ["AmplifyFoundation"]
+        ),
+        .library(
+            name: "AmplifyFoundationBridge",
+            targets: ["AmplifyFoundationBridge"]
         ),
     ],
     dependencies: dependencies,